Move lang_items definitions to librustc_lang_items.
